More reorg
- It compiles, so that's good I think
- Unsure about making the struct fields public, hmm
