+++ src/lib.rs
@@ -1,5 +1,5 @@
-pub mod online_coder;
-pub mod block_iter;
-pub mod decoder;
-pub mod types;
-
+pub mod encode;
+pub mod iter;
+pub mod decode;
+pub mod typedef;
+pub mod util;
